Fix issue #3
Some networks have public transport services with same 'ref' and 'route' in different cities/villages.
analyze_routes.pl must consider this by checking how many lines are in the CSV file with same 'ref' and 'route' values but different From and To and (!) Operator values.
In such cases, the code now tries to distinguish not only by the 'ref' and 'route' values but by also taking 'operator' into account.
Appropiare notes will be given to STDERR or in the WIKI file when this could not be done propperly.