fix issue #14
Some donwloads via Overpass-API query do not include all data that is references in the relations.
I.e. there are some routes with way members, but details for the ways (tags and node-IDs) are not in the downloaded data. This may also the case for member nodes.
The code depends on completeness of data and does not check like if ( $WAYS($way_id)->{'first_node'} )

Before analyzing route relations, check for completeness of data.
If not complete, issue an error in the analysis file (for this route) and on STDERR specifying the Relation-ID.
Furthermore skip check which depend in completeness of data.