fix issue #17
we already hava a special analysis for the first way being a oneway way.
extend this by analyzing whether the entry node of the first way (=oneway) corresponds to the first or last node of the second way.
if so, we know that the first way (=oneway) is used in wrong direction "PTv2 route: using oneway way(s) in wrong direction:".